refactor: refactor http server with typescript
